Replace int by uint where it's unnecessary to deal with negative numbers

+++ lib/server/install/step.go
@@ -267,7 +267,7 @@ func (is *step) Run(hosts []*pb.Host, v Variables, s Settings) (results StepResu
 	tracer := concurrency.NewTracer(is.Worker.feature.task, "", true).GoingIn()
 	defer tracer.OnExitTrace()()
 	defer scerr.OnExitLogError(tracer.TraceMessage(""), &err)()
-	nHosts := len(hosts)
+	nHosts := uint(len(hosts))
 	defer temporal.NewStopwatch().OnExitLogWithLevel(
 		fmt.Sprintf("Starting step '%s' on %d host%s...", is.Name, nHosts, utils.Plural(nHosts)),
 		fmt.Sprintf("Ending step '%s' on %d host%s", is.Name, len(hosts), utils.Plural(nHosts)),
